Make the project buildable and with up to date configuration (#1)
* Adding uncommitted files
* Switch the solution to Visual Studio 2022 and remove some autogenerated files
* Intermediate 1
* Intermediate 2
* Intermediate 3
* Intermediate 4
* Intermediate 5
* Intermediate 6
* Intermediate 7
* Update common submodule to recent existing
* Remove intermediate files
* Add README.md file
* Stick to a recent common
* Intermediate 8
* Intermediate 9
* Intermediate 10
* Intermediate 11
* Intermediate 12
* Intermediate 13 Move DirectPrint one level up
* Intermediate 13 Refactor LabBarcode project into a form of simple ATL project
* Intermediate 14 Refactor LabBarcode project done
* Intermediate 15 Remove outdated configuration
* Intermediate 15 Add REDBASE, ORCLE, and MSSQL common configuration
* Intermediate 16
* Intermediate 17
* Intermediate 18
* Intermediate 19
* Move SocketDll into the root directory
* Intermediate 19
* Intermediate 20
* Intermediate 21
* Intermediate 21
* Intermediate 22
* Intermediate 23
* Intermediate 25
* Intermediate 26
* Intermediate 27
* Intermediate 28
* Intermediate 29
* Intermediate 30
* Intermediate 31
* Intermediate 32
* Intermediate 33
* Update README.md to describe an issue building zLib
* Intermediate 34
* Intermediate 35
* Intermediate 36
* Intermediate 37
* Intermediate 38
* Intermediate 39
* Intermediate 40
* Intermediate 41
* Intermediate 42. Add xmlutil namespace
* Intermediate. Move Fprint-02K and Shtrih M idl files to common
* Update README.md to add a note about gSOAP.
* Intermediate 43. gSOAP enable
* Intermediate 44. Use Boost 1.87
* Intermediate 45. Update configuration to remove unused targets
* Intermediate 46
* Intermediate 47
* Intermediate 48
* Intermediate 49
* Intermediate 50
* Intermediate 51
* Intermediate 51. Done, zero compilation errors
* Intermediate 52. Warnings
* Intermediate 53. Warnings return paths
* Intermediate 54. Warnings return paths
* Intermediate 55. GUIDs
* Intermediate 56. UNICODE
* Intermediate 56. ORACLE debug
* Intermediate 56. Redbase release
* Intermediate 57. Oracle release
* Intermediate 57. Remove useless configuration stuff
* Intermediate 58. Move Test project one level up
* Intermediate 59. All targets are buildable!
* Intermediate 60. WTL 10
* Intermediate 61. Fix visual issues in test application
* Intermediate 62. Fix minor issues in test application
* Intermediate 63. INI stuff
* Intermediate 64. Typo.
* Intermediate 65. WINCE removed, missed files added, a project moved to another location, working on configuration.
* Intermediate 66, working on configuration.
* Intermediate 67, working on configuration.
* Intermediate 68, working on configuration.
* Intermediate 69, working on configuration.
* Intermediate 70, working on configuration.
* Intermediate 71, working on configuration.
* Intermediate 72, working on configuration.
* Intermediate 73, working on configuration.
* Intermediate 74, working on configuration.
* Intermediate 75, working on configuration.
* Intermediate 76, working on configuration. Enable JUMBO build for some projects
* Intermediate 77, working on configuration.
* Intermediate 78, working on configuration.
* Intermediate 79, JUMBO build
* Intermediate 79, Warnings
* Intermediate 80. Change MIDL generated file names to more standard.
* Intermediate 81. Normalize everything to UTF-8 encoded source files
* Updated common; All paths in the projects are normalized by [MSBuild]::NormalizePath; Removed legacy code from the configuration.
* Intermediate. Some warnings were fixed, one promoted to an error. Minor issues fixed. Deployment of external dependencies now part of the build.
* Fix release build
* Alpha. Fixed clear defects detected during the first run. Typos. Minor code issues. Import directive problem. Get parameters rom DB issue.
* Alpha. Minor code issues.
* Alpha 2. Get rid of ExecutablePath settings in all projects. Remove AdditionalOptions in CL settings for some project. remove useless message box in ExportZv project
* Alpha 3. Get back correct AdditionalOptions in CL settings for some project that use #import directive.
* Add helpers to create process
* Add formatter project for formatting MFCCommon directory
* Alpha 4. Remove legacy AboutBox stuff from some projects; Replace CreateProcess API call by internal create_process implementation; Minor changes.
* Remove some memory leaks from AllocSysString call; Fix an issue with system_api::execute_process
* Spellchecking Russian text in all RC files and in some source code files
* Get rid of the License dialog stuff as it was implemented in the common laboratory.dll
* Saving some big bitmaps with lower bitness keeping them visually intact
* Update LabManager in order to work with recent splash
* Restoring memory tracking diagnostic defines to be able to track memory leaks; Recent version of splash screen stuff update
* Restoring memory tracking diagnostic defines to be able to track memory leaks; Spell checking and trivial changes.
* Restoring memory tracking defines and fixing typos in MyObject (#2)
* Delete unnecessary files
* Refactor MyObject, spell check, etc.
* Delete tree items on Reference close. Memory leak hunting.
* Add a first unit test project. Spell check SAHADIN.
* Add comments to the first unit test
* Add comments to the first unit test
* Add missed pragma once preprocessor directive for some files
* Rename props files in order to simplify the configuration
* Complete simplifying the configuration
* Get rid of IMP_EXP Aist stuff as it has only MFC skeleton projects
* Merge proxy/stub code into dlls. Update dlldata.c file name across all projects.
* Updating, fix minor issues.
* Fixes
* Fixes
* Get rid of the "Fluent UI" sample project comments
* Get rid of the "Fluent UI" sample project comments
* Backport from laboratory2 (typos)
* Intermediate. Minor changes.
* Update common submodule
* Upgrade SLN file to new SLNX format
* Clean-up projects. Add the SLNX file.
* Clear the buffer before use

+++ NewCellTypes/GridCellCheck.cpp
@@ -59,7 +59,7 @@ CSize CGridCellCheck::GetCellExtent(CDC* pDC)
     int nWidth = GetSystemMetrics(SM_CXHSCROLL) + 2*GetMargin();	
     CSize	cellSize = CGridCell::GetCellExtent(pDC);	
     cellSize.cx += nWidth;	
-    cellSize.cy = max (cellSize.cy, nWidth);	
+    cellSize.cy = max(cellSize.cy, (LONG)nWidth);
     return  cellSize;
 }
